Start working on annotation interface
