+++ Assets/Plugins/ARUI/Scripts/Annotation/Annotation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 290ea7888de2e124cb300b9face159a5
+guid: 09f02702d29b23944b5423e80928356b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
