--- Assets/Plugins/ARUI/Scripts/Annotation/ObjectAnchor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 09f02702d29b23944b5423e80928356b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
